+++ src/args.rs
@@ -1,8 +1,27 @@
 //! Definition of the available application arguments.
 use clap::Parser;
 
+const AFTER_HELP: &str = "
+COLORS
+
+    Whenever a color needs to be specified in the configuration, the
+    following three ways of specifying it are supported:
+
+     - One of the following words specifying a color from the 16 colors set:
+       black, dark_grey, red, dark_red, green, dark_green,
+       yellow, dark_yellow, blue, dark_blue, magenta,
+       dark_magenta, cyan, dark_cyan, white, grey.
+     - A color from the 256 colors set in format \"5;<color_index>\", e.g.
+       \"5;202\" for orange.
+     - An RGB color in format \"2;<r>;<g>;<b>\", e.g. \"2;255;255;0\" for
+       yellow.
+
+     Note that the ability to render different colors depends on your
+     terminal emulator and configuration.
+";
+
 #[derive(Debug, Parser)]
-#[command(author, version, about)]
+#[command(author, version, about, after_help=AFTER_HELP)]
 pub struct Args {
     /// File to select the text from. Omit to use standard input.
     pub file: Option<std::path::PathBuf>,