+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - The errors will now appear as tmux messages
 - `Change` Perform selection from the first page instead of the last when text longer than the screen is piped into mless
 - `Feature` It is now possible to define multiple modes in the config file and switch between them by pressing `<Space>` followed by the appropriate hotkey
+- `Feature` Colors of hints during selection can now be customized in the config file
 
 # v0.1.0
 First release, no changelog was kept